+++ b2/resource_b2_application_key_test.go
@@ -37,17 +37,19 @@ func TestAccResourceB2ApplicationKey_basic(t *testing.T) {
 					resource.TestCheckResourceAttr(resourceName, "bucket_id", ""), // deprecated
 					resource.TestCheckResourceAttr(resourceName, "capabilities.#", "1"),
 					resource.TestCheckResourceAttr(resourceName, "capabilities.0", "readFiles"),
+					resource.TestCheckResourceAttr(resourceName, "expiration_timestamp", "0"),
 					resource.TestCheckResourceAttr(resourceName, "key_name", keyName),
 					resource.TestCheckResourceAttr(resourceName, "name_prefix", ""),
 					resource.TestCheckResourceAttr(resourceName, "options.#", "1"),
 					resource.TestCheckResourceAttr(resourceName, "options.0", "s3"),
+					resource.TestCheckResourceAttr(resourceName, "valid_duration_in_seconds", "0"),
 				),
 			},
 			{
 				ResourceName:            resourceName,
 				ImportState:             true,
 				ImportStateVerify:       true,
-				ImportStateVerifyIgnore: []string{"application_key"},
+				ImportStateVerifyIgnore: []string{"application_key", "valid_duration_in_seconds"},
 			},
 		},
 	})
@@ -73,17 +75,19 @@ func TestAccResourceB2ApplicationKey_all(t *testing.T) {
 					resource.TestCheckResourceAttrPair(resourceName, "bucket_id", parentResourceName, "bucket_id"), // deprecated
 					resource.TestCheckResourceAttr(resourceName, "capabilities.#", "1"),
 					resource.TestCheckResourceAttr(resourceName, "capabilities.0", "writeFiles"),
+					resource.TestMatchResourceAttr(resourceName, "expiration_timestamp", regexp.MustCompile("^[1-9][0-9]{12,}$")),
 					resource.TestCheckResourceAttr(resourceName, "key_name", keyName),
 					resource.TestCheckResourceAttr(resourceName, "name_prefix", "prefix"),
 					resource.TestCheckResourceAttr(resourceName, "options.#", "1"),
 					resource.TestCheckResourceAttr(resourceName, "options.0", "s3"),
+					resource.TestCheckResourceAttr(resourceName, "valid_duration_in_seconds", "86400"),
 				),
 			},
 			{
 				ResourceName:            resourceName,
 				ImportState:             true,
 				ImportStateVerify:       true,
-				ImportStateVerifyIgnore: []string{"application_key"},
+				ImportStateVerifyIgnore: []string{"application_key", "valid_duration_in_seconds"},
 			},
 		},
 	})
@@ -109,10 +113,12 @@ func TestAccResourceB2ApplicationKey_deprecatedBucketId(t *testing.T) {
 					resource.TestCheckResourceAttrPair(resourceName, "bucket_id", parentResourceName, "bucket_id"), // deprecated
 					resource.TestCheckResourceAttr(resourceName, "capabilities.#", "1"),
 					resource.TestCheckResourceAttr(resourceName, "capabilities.0", "writeFiles"),
+					resource.TestMatchResourceAttr(resourceName, "expiration_timestamp", regexp.MustCompile("^[1-9][0-9]{12,}$")),
 					resource.TestCheckResourceAttr(resourceName, "key_name", keyName),
 					resource.TestCheckResourceAttr(resourceName, "name_prefix", "prefix"),
 					resource.TestCheckResourceAttr(resourceName, "options.#", "1"),
 					resource.TestCheckResourceAttr(resourceName, "options.0", "s3"),
+					resource.TestCheckResourceAttr(resourceName, "valid_duration_in_seconds", "86400"),
 				),
 			},
 			{
@@ -125,10 +131,12 @@ func TestAccResourceB2ApplicationKey_deprecatedBucketId(t *testing.T) {
 					resource.TestCheckResourceAttrPair(resourceName, "bucket_id", parentResourceName, "bucket_id"), // deprecated
 					resource.TestCheckResourceAttr(resourceName, "capabilities.#", "1"),
 					resource.TestCheckResourceAttr(resourceName, "capabilities.0", "writeFiles"),
+					resource.TestMatchResourceAttr(resourceName, "expiration_timestamp", regexp.MustCompile("^[1-9][0-9]{12,}$")),
 					resource.TestCheckResourceAttr(resourceName, "key_name", keyName),
 					resource.TestCheckResourceAttr(resourceName, "name_prefix", "prefix"),
 					resource.TestCheckResourceAttr(resourceName, "options.#", "1"),
 					resource.TestCheckResourceAttr(resourceName, "options.0", "s3"),
+					resource.TestCheckResourceAttr(resourceName, "valid_duration_in_seconds", "86400"),
 				),
 			},
 		},
@@ -156,6 +164,7 @@ resource "b2_application_key" "test" {
   capabilities = ["writeFiles"]
   bucket_ids = [b2_bucket.test.bucket_id]
   name_prefix = "prefix"
+  valid_duration_in_seconds = 86400
 }
 `, keyName, keyName)
 }
@@ -172,6 +181,7 @@ resource "b2_application_key" "test" {
   capabilities = ["writeFiles"]
   bucket_id = b2_bucket.test.bucket_id
   name_prefix = "prefix"
+  valid_duration_in_seconds = 86400
 }
 `, keyName, keyName)
 }