+++ CHANGELOG.md
@@ -6,14 +6,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+* Add `expiration_timestamp` to `b2_application_key` resource and data source
+* Add `valid_duration_in_seconds` to `b2_application_key` resource
+
 ## [0.12.0] - 2025-12-04
 
 ### Added
 * Add `buckets` list to `b2_account_info` data source to support multi-bucket application keys
 * Add `bucket_ids` list to `b2_application_key` resource and data source to support multi-bucket application keys
 * Add `recommended_part_size` and `absolute_minimum_part_size` to `b2_account_info` data source
-* Add `expiration_timestamp` to `b2_application_key` resource and data source
-* Add `valid_duration_in_seconds` to `b2_application_key` resource
 
 ### Changed
 * Migrate to B2 native api v4