+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+* Use B2 Native API v3 when using deprecated `bucket_id`
+
 ## [0.12.0] - 2025-12-04
 
 ### Added