+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.12.1] - 2026-02-21
+
 ### Fixed
 * Use B2 Native API v3 when using deprecated `bucket_id`
 
@@ -245,7 +247,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 * Implementation of PoC (simple `b2_application_key` data source)
 
-[Unreleased]: https://github.com/Backblaze/terraform-provider-b2/compare/v0.12.0...HEAD
+[Unreleased]: https://github.com/Backblaze/terraform-provider-b2/compare/v0.12.1...HEAD
+[0.12.1]: https://github.com/Backblaze/terraform-provider-b2/compare/v0.12.0...v0.12.1
 [0.12.0]: https://github.com/Backblaze/terraform-provider-b2/compare/v0.11.0...v0.12.0
 [0.11.0]: https://github.com/Backblaze/terraform-provider-b2/compare/v0.10.0...v0.11.0
 [0.10.0]: https://github.com/Backblaze/terraform-provider-b2/compare/v0.9.0...v0.10.0